w3c-reports/12092014/mac/runned-tests/IndexedDB/idbfactory_open3-expected.txt

CONSOLE MESSAGE: line 69: TypeError: null is not an object (evaluating 'window.indexedDB.open')

Harness Error. harness_status.status = 1 , harness_status.message = TypeError: null is not an object (evaluating 'window.indexedDB.open')
